PatchDay 1.1
Changed the name of the “to” button on the Patch Details View to “Change Patch”.
The general locations picker is activated when tapping the location now, and for custom locations, you tap the keyboard button to the right (on Patch Details View).
The “X” button in the top corner of Patch Details View in more aligned
The headings on the Schedule View is more aligned.
Removed the extra button for editing the patch date.
The expiration date attribute is now more readable and part of the stack in the Patch Details View.
The expiration date changes when the patch’s date-placed changes in the Patch Details View.
The day of the week was added to the dates.
All dates have the same format, and it works accordingly to the user’s settings for dates.
On the Schedule View, there is now a button titled “Next Patch” that displays an alert telling the user when and where the next patch to change is.
Fixed bug where the app crashes if there is no set notification set on record and the “Suggest Patch” functionality is turned off, and the save button is hit.
Fixed other bugs